update 4-3-2025 part 2

+++ app/Repositories/BantuanRepository.php
@@ -68,7 +68,7 @@ class BantuanRepository
                             ->where('jumlah_bantuan', $kelompokTani->luas_lahan * 15)->first();
                         if ($cekBantuan) {
                             $bantuan = $cekBantuan;
-                        }else{
+                        } else {
                             $bantuan = new Bantuan();
                             $bantuan->id_kab_kota = $data['id_kab_kota'];
                             $bantuan->jenis_bantuan = $data['jenis_bantuan'];
@@ -78,14 +78,12 @@ class BantuanRepository
                             $bantuan->tahun = $data['tahun'];
                             // $bantuan->keterangan = $data['keterangan'];
                             $bantuan->user_id = $data['user_id'];
-                            $bantuan->save();       
-                        }            
+                            $bantuan->save();
+                        }
                         BantuanKelompokTani::create([
                             'bantuan_id' => $bantuan->id,
                             'kelompok_tani_id' => $kelompokTaniId,
                         ]);
-                        
-            
                     }
                 }
                 DB::commit();
@@ -96,7 +94,6 @@ class BantuanRepository
                     ]
                 ];
             }
-            
         } catch (\Exception $e) {
             DB::rollBack();
             return [
@@ -111,30 +108,49 @@ class BantuanRepository
         DB::beginTransaction();
         try {
             $bantuan = $this->bantuanModel->find($id);
-            $bantuan->id_kab_kota = $data['id_kab_kota'];
-            $bantuan->jenis_bantuan = $data['jenis_bantuan'];
-            $bantuan->jumlah_bantuan = $data['jumlah_bantuan'];
-            $bantuan->satuan_bantuan = $data['satuan_bantuan'];
-            $bantuan->bulan = $data['bulan'];
-            $bantuan->tahun = $data['tahun'];
-            // $bantuan->keterangan = $data['keterangan'];
-            $bantuan->save();
+            if (!$bantuan) {
+                return [
+                    "id" => '0',
+                    "data" => "Bantuan not found"
+                ];
+            }
 
             if (isset($data['kelompok_tani_id']) && is_array($data['kelompok_tani_id'])) {
                 BantuanKelompokTani::where('bantuan_id', $id)->delete();
-                foreach ($data['kelompok_tani_id'] as $kelompokTaniId) {
-                    BantuanKelompokTani::create([
-                        'bantuan_id' => $id,
-                        'kelompok_tani_id' => $kelompokTaniId,
-                    ]);
-                }
-            }
 
-            DB::commit();
-            return [
-                "id" => '1',
-                "data" => 'update data bantuan success'
-            ];
+                foreach ($data['kelompok_tani_id'] as $kelompokTaniId) {
+                    $kelompokTani = KelompokTani::find($kelompokTaniId);
+                    if ($kelompokTani) {
+                        $cekBantuan = Bantuan::where('jenis_bantuan', $data['jenis_bantuan'])
+                            ->where('tahun', $data['tahun'])
+                            ->where('bulan', $data['bulan'])
+                            ->where('jumlah_bantuan', $kelompokTani->luas_lahan * 15)
+                            ->first();
+
+                        if ($cekBantuan) {
+                            $bantuan = $cekBantuan;
+                        } else {
+                            $bantuan->id_kab_kota = $data['id_kab_kota'];
+                            $bantuan->jenis_bantuan = $data['jenis_bantuan'];
+                            $bantuan->jumlah_bantuan = $kelompokTani->luas_lahan * 15; // Recalculate
+                            $bantuan->satuan_bantuan = $data['satuan_bantuan'];
+                            $bantuan->bulan = $data['bulan'];
+                            $bantuan->tahun = $data['tahun'];
+                            $bantuan->save();
+                        }
+
+                        BantuanKelompokTani::create([
+                            'bantuan_id' => $bantuan->id,
+                            'kelompok_tani_id' => $kelompokTaniId,
+                        ]);
+                    }
+                }
+                DB::commit();
+                return [
+                    "id" => '1',
+                    "data" => "Update data bantuan success"
+                ];
+            }
         } catch (\Exception $e) {
             DB::rollBack();
             return [
